test(app): isolate local state helpers from router context
