+++ packages/app/src/context/local.tsx
@@ -8,17 +8,9 @@ import { useProviders } from "@/hooks/use-providers"
 import { modelEnabled, modelProbe } from "@/testing/model-selection"
 import { Persist, persisted } from "@/utils/persist"
 import { cycleModelVariant, getConfiguredAgentVariant, resolveModelVariant } from "./model-variant"
+import { clone, getAgentKey as key, pickAgentItem, syncSessionState, type ModelKey, type State } from "./local-state"
 import { useSDK } from "./sdk"
 import { useSync } from "./sync"
-
-export type ModelKey = { providerID: string; modelID: string }
-
-type State = {
-  agent?: string
-  model?: ModelKey
-  variant?: string | null
-  message?: string
-}
 
 type Saved = {
   session: Record<string, State | undefined>
@@ -28,31 +20,6 @@ const WORKSPACE_KEY = "__workspace__"
 const handoff = new Map<string, State>()
 
 const handoffKey = (dir: string, id: string) => `${dir}\n${id}`
-
-const key = (item: { id?: string; name: string }) => item.id ?? item.name
-
-const pickByKey = <T extends { id?: string; name: string }>(items: T[], value: string | undefined) => {
-  if (!value) return undefined
-  return items.find((item) => key(item) === value) ?? items.find((item) => item.name === value)
-}
-
-export const pickAgentItem = <T extends { id?: string; name: string }>(items: T[], value: string | undefined) => {
-  if (items.length === 0) return undefined
-  return pickByKey(items, value) ?? items[0]
-}
-
-export const syncSessionState = (
-  prev: State | undefined,
-  msg: { id: string; agent: string; model: ModelKey; variant?: string },
-) => {
-  if (prev?.message === msg.id) return
-  return {
-    agent: msg.agent,
-    model: msg.model,
-    variant: msg.variant ?? null,
-    message: msg.id,
-  } satisfies State
-}
 
 const migrate = (value: unknown) => {
   if (!value || typeof value !== "object") return { session: {} }
@@ -68,14 +35,6 @@ const migrate = (value: unknown) => {
   return {
     session: Object.fromEntries(Object.entries(item.pick).filter(([key]) => key !== WORKSPACE_KEY)),
   }
-}
-
-const clone = (value: State | undefined) => {
-  if (!value) return undefined
-  return {
-    ...value,
-    model: value.model ? { ...value.model } : undefined,
-  } satisfies State
 }
 
 export const { use: useLocal, provider: LocalProvider } = createSimpleContext({
